backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
yfinance==0.2.32
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.2
joblib==1.3.2
requests==2.31.0
feedparser==6.0.10
python-telegram-bot==20.3
twilio==8.10.0
schedule==1.2.0
pytz==2023.3
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
scipy==1.11.4
aiofiles==23.2.1
vaderSentiment==3.3.2
transformers==4.34.0
torch==2.1.0
